+++ protractor.config.js
@@ -1,9 +1,14 @@
-// TO RUN THE TESTS
-//
-// The first time, run:
+// FIRST TIME ONLY- run:
 //   ./node_modules/.bin/webdriver-manager update
-// Make sure the test server is running. Then do.
-//   ./node_modules/.bin/protractor protractor.config.js
+//
+//   Try: `npm run webdriver:update`
+//
+// AND THEN EVERYTIME ...
+//   1. Compile with `tsc`
+//   2. Make sure the test server (e.g., http-server: localhost:8080) is running.
+//   3. ./node_modules/.bin/protractor protractor.config.js
+//
+// To do all steps, try:  `npm run e2e`
 
 var fs = require('fs');
 var path = require('canonical-path');